requirements.txt
asgiref==3.9.0
attrs==25.3.0
channels==4.2.2
channels_redis==4.2.1
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
msgpack==1.1.1
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
rpds-py==0.26.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
